requirement.txt
astroid==1.6.4
Babel==2.6.0
dataclasses==0.6
Django==2.1
django-bootstrap4==0.0.7
django-phonenumber-field==2.0.0
djongo==1.2.29
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
phonenumberslite==8.9.10
pkg-resources==0.0.0
pylint==1.9.1
pylint-django==0.11.1
pylint-plugin-utils==0.2.6
pymongo==3.6.1
pytz==2018.4
six==1.11.0
sqlparse==0.2.4
wrapt==1.10.11
